Use slower (but more cross-platform) method of invoking shell commands
Different unixen have different maximum lengths for command line so when
`cp * /some/place` was expanded by the shell it would exceed that limit.

From now we iterate over files in dirs and invokes a separate instance of
the command for each file - this is slower but more reliable.

This commit also makes the script more verbose in its output so the user
can track progress more easily

+++ build-assets.py
@@ -7,6 +7,9 @@ import shutil
 
 import freelex
 
+def print_run_msg(msg):
+    print(" - Running: " + msg)
+
 parser = OptionParser()
 parser.add_option("-i", "--ios",
                   help="location of iOS app root", metavar="IOS_PATH")
@@ -63,50 +66,93 @@ for path, dirs, files in os.walk("assets/"):
     for filename in files:
         # Some images have a 1px border that looks bad in search results
         # Not all do - but we can safely trim 1px off all images
-        os.system("mogrify -shave 1x1 assets/" + filename)
+        shave_cmd = "mogrify -shave 1x1 assets/" + filename
+        print_run_msg(shave_cmd)
+        os.system(shave_cmd)
 
         # Then we make thumbnails of the border-free images
-        os.system("convert -resize x92 assets/" + filename + " assets/50." + filename)
+        create_thumbnail_cmd = "convert -resize x92 assets/" + filename + " assets/50." + filename
+        print_run_msg(create_thumbnail_cmd)
+        os.system(create_thumbnail_cmd)
 
 print("Step 6p: Shrink images for distribution")
 # In order to keep the app size small, we need to run a series
 # of compressions over the images
 
 # Resize images larger than 600x600 down using mogrify from imagemagick
-os.system("mogrify -resize '600x600>' assets/*.png")
+for path, dirs, files in os.walk("assets/"):
+    for filename in files:
+        cmd = "mogrify -resize '600x600>' assets/" + filename
+        print_run_msg(cmd)
+        os.system(cmd)
 
 # Convert all images to 4 colour depth
 for path, dirs, files in os.walk("assets/"):
     for filename in files:
-        os.system("convert -colors 4 assets/" + filename + " assets/" + filename)
+        recolor_cmd = "convert -colors 4 assets/" + filename + " assets/" + filename
+        print_run_msg(recolor_cmdcmd)
+        os.system(recolor_cmd)
 
 # Finally, run optipng
-os.system("optipng assets/*.png")
+for path, dirs, files in os.walk("assets/"):
+    for filename in files:
+        optipng_cmd = "optipng -quiet assets/" + filename
+        print_run_msg(optipng_cmd)
+        os.system(optipng_cmd)
 
 print("Step 7a: Update iOS app images")
+
 ios_asset_path = options.ios + "/Data/picture/"
-shutil.rmtree(ios_asset_path)
+
+if os.path.isdir(ios_asset_path):
+    shutil.rmtree(ios_asset_path)
+
 os.makedirs(ios_asset_path)
-os.system("cp assets/* " + ios_asset_path)
+
+# re-create the .gitkeep file in the assets dir (it is not necessary for the
+# app but keeps that dir in the git repo)
+os.system("touch " + android_asset_path + ".gitkeep")
+
+for path, dirs, files in os.walk("assets/"):
+    for filename in files:
+        cp_cmd = "cp assets/" + filename + " " + ios_asset_path
+        print_run_msg(cp_cmd)
+        os.system(cp_cmd)
 
 print("Step 7b: Update iOS app nzsl.db")
 os.system("cp nzsl.db " + options.ios + "/Data/")
 
 print("Step 8a: Update Android app images")
 android_asset_path = options.android + "/app/src/main/assets/images/signs/"
-shutil.rmtree(android_asset_path)
+
+if os.path.isdir(android_asset_path):
+    shutil.rmtree(android_asset_path)
+
 os.makedirs(android_asset_path)
-os.system("cp assets/* " + android_asset_path)
+
+# re-create the .gitkeep file in the assets dir (it is not necessary for the
+# app but keeps that dir in the git repo)
+os.system("touch " + android_asset_path + ".gitkeep")
+
+for path, dirs, files in os.walk("assets/"):
+    for filename in files:
+        cp_cmd = "cp assets/" + filename + " " + android_asset_path
+        print_run_msg(cp_cmd)
+        os.system(cp_cmd)
 
 print("Step 8b: Update Android app nzsl.dat")
-os.system("cp nzsl.dat " + options.android + "/app/src/main/assets/db/")
+android_db_path = options.android + "/app/src/main/assets/db/"
+os.makedirs(android_db_path)
+os.system("cp nzsl.dat " + android_db_path)
 
-if (options.cleanup == True):
+if options.cleanup:
     print("Step 9: Cleanup")
     os.remove("dnzsl-xmldump.xml")
     os.remove("nzsl.dat")
     os.remove("nzsl.db")
     shutil.rmtree("picture")
     shutil.rmtree("assets")
+else:
+    print("Skipping cleanup (see --help for how to enable it)")
 
 print("Done")
